Move RequirementCommand to req_command.py.
