+++ src/pip/_internal/commands/install.py
@@ -11,8 +11,8 @@ from pip._vendor import pkg_resources
 
 from pip._internal.cache import WheelCache
 from pip._internal.cli import cmdoptions
-from pip._internal.cli.base_command import RequirementCommand
 from pip._internal.cli.cmdoptions import make_target_python
+from pip._internal.cli.req_command import RequirementCommand
 from pip._internal.cli.status_codes import ERROR
 from pip._internal.exceptions import (
     CommandError,
